Update zh_CN.lang file ------ version 1.0 pre 5.5
I have a few questions about the lang file. For line 57 and 59 (in en_US.lang), it seems that they are the names of the same planet but somehow have different names (Oberon/Oberón)? I don't know if this is a mistake or it is intended. Similar things happen between line 820-825 and line 827-832 (I delete these in zh_CN.lang) which seems to be the descriptions of the same achievements. If it is intended plz don't merge and I'll make some corrections.